[DP-328] refactor: hooks 폴더 구조 분리 #332 (#353)
* [DP-328] refactor: hooks 폴더 구조 분리 #332

* [DP-328] fix: fix build error

---------
