+++ src/feature/map/components/pages/MapSelectLocation.tsx
@@ -5,7 +5,7 @@ import { AnimatePresence, motion } from "framer-motion";
 import InteractiveMap from "@/feature/map/components/sections/regist/InteractiveMap";
 import MapHeaderSearchBox from "@/feature/map/components/sections/regist/MapHeaderSearchBox";
 import MapSelectedInfoCard from "@/feature/map/components/sections/regist/MapSelectedInfoCard";
-import { useMapLocationSelector } from "@/feature/map/hooks/useMapLocationSelector";
+import { useMapLocationSelector } from "@feature/map/hooks/map/useMapLocationSelector";
 
 function typeGuard(target: unknown): target is "hotspot" | "wifi" {
   if (target === "hotspot" || target === "wifi") return true;
