+++ src/feature/map/hooks/register/useRegisterFormValidation.ts
